requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pymongo==4.6.0
motor==3.3.1
python-dotenv==1.0.0
httpx==0.25.1
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
openai==0.28.1
pytest==7.4.3
loguru==0.7.2
lxml==4.9.3 